+++ packages/eyes-cypress/src/plugin/hooks.js
@@ -4,40 +4,39 @@ const makeHandleBatchResultsFile = require('./makeHandleBatchResultsFile');
 const getErrorsAndDiffs = require('./getErrorsAndDiffs');
 const processCloseAndAbort = require('./processCloseAndAbort');
 const errorDigest = require('./errorDigest');
-const {tests} = require('./runningTests');
+const runningTests = require('./runningTests');
 
-function setGlobalRunHooks(on, {visualGridClient, logger}) {
+function makeGlobalRunHooks({visualGridClient, logger}) {
   let waitForBatch;
 
-  on('before:run', ({config}) => {
-    const {isTextTerminal, eyesTestConcurrency: testConcurrency} = config;
-    // ugly but neccessary here - as this is the only place that cypress exposes the run mode on config
-    if (!isTextTerminal) return;
+  return {
+    'before:run': ({config}) => {
+      const {isTextTerminal, eyesTestConcurrency: testConcurrency} = config;
+      if (!config.isTextTerminal) return;
 
-    waitForBatch = makeWaitForBatch({
-      logger: (logger.extend && logger.extend('waitForBatch')) || console,
-      testConcurrency,
-      processCloseAndAbort,
-      getErrorsAndDiffs,
-      errorDigest,
-      isInteractive: !isTextTerminal,
-      handleBatchResultsFile: makeHandleBatchResultsFile(config),
-    });
-  });
+      waitForBatch = makeWaitForBatch({
+        logger: (logger.extend && logger.extend('waitForBatch')) || console,
+        testConcurrency,
+        processCloseAndAbort,
+        getErrorsAndDiffs,
+        errorDigest,
+        isInteractive: !isTextTerminal,
+        handleBatchResultsFile: makeHandleBatchResultsFile(config),
+      });
+    },
 
-  on('after:run', async ({config}) => {
-    if (!config.isTextTerminal) return;
+    'after:run': async ({config}) => {
+      if (!config.isTextTerminal) return;
 
-    try {
-      // we rely on the fact that `runningTests.reset()` wasn't called
-      // so that tests that were added in `open` are available here (through `tests` in `runningTests.js`)
-      await waitForBatch(tests, visualGridClient.closeBatch);
-    } catch (e) {
-      if (!!config.eyesFailCypressOnDiff) {
-        throw e;
+      try {
+        await waitForBatch(runningTests.tests, visualGridClient.closeBatch);
+      } catch (e) {
+        if (!!config.eyesFailCypressOnDiff) {
+          throw e;
+        }
       }
-    }
-  });
+    },
+  };
 }
 
-module.exports = setGlobalRunHooks;
+module.exports = makeGlobalRunHooks;