+++ packages/eyes-cypress/src/plugin/pluginExport.js
@@ -1,22 +1,38 @@
 'use strict';
-const setGlobalRunHooks = require('./hooks');
 const shouldSetGlobalHooks = require('./shouldSetGlobalHooks');
+const {presult} = require('@applitools/functional-commons');
 
-function makePluginExport({startServer, eyesConfig, visualGridClient, logger}) {
+function makePluginExport({startServer, eyesConfig, globalHooks}) {
   return function pluginExport(pluginModule) {
     let closeEyesServer;
     const pluginModuleExports = pluginModule.exports;
     pluginModule.exports = async (...args) => {
       const {eyesPort, closeServer} = await startServer();
       closeEyesServer = closeServer;
-      const moduleExportsResult = await pluginModuleExports(...args);
-      const [on, config] = args;
-
-      if (shouldSetGlobalHooks({...config, ...eyesConfig})) {
-        setGlobalRunHooks(on, {visualGridClient, logger, eyesConfig});
-      }
-
+      const [origOn, config] = args;
+      const moduleExportsResult = await pluginModuleExports(onThatCallsUserDefinedHandler, config);
       return Object.assign({}, eyesConfig, {eyesPort}, moduleExportsResult);
+
+      // This piece of code exists because at the point of writing, Cypress does not support multiple event handlers:
+      // https://github.com/cypress-io/cypress/issues/5240#issuecomment-948277554
+      // So we wrap Cypress' `on` function in order to wrap the user-defined handler. This way we can call our own handler
+      // in addition to the user's handler
+      function onThatCallsUserDefinedHandler(eventName, handler) {
+        const isRunEvent = eventName === 'before:run' || eventName === 'after:run';
+        const handlerToCall =
+          shouldSetGlobalHooks(config) && isRunEvent ? handlerThatCallsUserDefinedHandler : handler;
+        return origOn.call(this, eventName, handlerToCall);
+
+        async function handlerThatCallsUserDefinedHandler() {
+          const [err] = await presult(
+            Promise.resolve(globalHooks[eventName].apply(this, arguments)),
+          );
+          await handler.apply(this, arguments);
+          if (err) {
+            throw err;
+          }
+        }
+      }
     };
     return function getCloseServer() {
       return closeEyesServer;