+++ packages/eyes-cypress/test/fixtures/testApp/cypress/integration-run/global-hooks-overrides.js
@@ -1,0 +1,5 @@
+describe('eyes-cypress', () => {
+  it('works', () => {
+    expect(true).to.be.true;
+  });
+});